refactor: 일기 반응 도메인 분리 TODAK-249
- 공개 일기에서 반응 도메인을 분리(Controller, Service)
