+++ src/test/java/com/heartsave/todaktodak_api/diary/controller/PublicDiaryControllerTest.java
@@ -1,6 +1,5 @@
 package com.heartsave.todaktodak_api.diary.controller;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
@@ -13,9 +12,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.heartsave.todaktodak_api.common.security.WithMockTodakUser;
-import com.heartsave.todaktodak_api.diary.constant.DiaryReactionType;
 import com.heartsave.todaktodak_api.diary.dto.PublicDiary;
-import com.heartsave.todaktodak_api.diary.dto.request.PublicDiaryReactionRequest;
 import com.heartsave.todaktodak_api.diary.dto.request.PublicDiaryWriteRequest;
 import com.heartsave.todaktodak_api.diary.dto.response.PublicDiaryPaginationResponse;
 import com.heartsave.todaktodak_api.diary.service.PublicDiaryService;
@@ -67,73 +64,6 @@ public class PublicDiaryControllerTest {
             post("/api/v1/diary/public")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(request)))
-        .andExpect(status().isBadRequest());
-  }
-
-  @Test
-  @DisplayName("정상적인 반응 토글 요청 시 성공")
-  @WithMockTodakUser
-  void toggleReaction_Success() throws Exception {
-    Long diaryId = 1L;
-    DiaryReactionType reactionType = DiaryReactionType.LIKE;
-    PublicDiaryReactionRequest request = new PublicDiaryReactionRequest(diaryId, reactionType);
-
-    doNothing()
-        .when(publicDiaryService)
-        .toggleReactionStatus(anyLong(), any(PublicDiaryReactionRequest.class));
-
-    mockMvc
-        .perform(
-            post("/api/v1/diary/public/reaction")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)))
-        .andExpect(status().isNoContent());
-  }
-
-  @Test
-  @DisplayName("반응 토글 실패 - 잘못된 일기 ID")
-  @WithMockTodakUser
-  void toggleReaction_Fail_InvalidDiaryId() throws Exception {
-    Long invalidDiaryId = 0L;
-    DiaryReactionType reactionType = DiaryReactionType.LIKE;
-    PublicDiaryReactionRequest request =
-        new PublicDiaryReactionRequest(invalidDiaryId, reactionType);
-
-    mockMvc
-        .perform(
-            post("/api/v1/diary/public/reaction")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)))
-        .andExpect(status().isBadRequest());
-  }
-
-  @Test
-  @DisplayName("반응 토글 실패 - 반응 타입이 null")
-  @WithMockTodakUser
-  void toggleReaction_Fail_NullReactionType() throws Exception {
-    Long diaryId = 1L;
-    DiaryReactionType reactionType = null;
-    PublicDiaryReactionRequest request = new PublicDiaryReactionRequest(diaryId, reactionType);
-
-    mockMvc
-        .perform(
-            post("/api/v1/diary/public/reaction")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)))
-        .andExpect(status().isBadRequest());
-  }
-
-  @Test
-  @DisplayName("반응 토글 실패 - 잘못된 요청 형식")
-  @WithMockTodakUser
-  void toggleReaction_Fail_InvalidRequestFormat() throws Exception {
-    String invalidRequest = "{\"publicDiaryId\": \"invalid\", \"reactionType\": \"like\"}";
-
-    mockMvc
-        .perform(
-            post("/api/v1/diary/public/reaction")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(invalidRequest))
         .andExpect(status().isBadRequest());
   }
